owner validation on PUT/DELETE routes

+++ app/api/menu_items.py
@@ -1,31 +1,32 @@
 from flask import Blueprint, jsonify, request
-from flask_login import login_required
-from ..models.menu_item import MenuItem, db
-from ..forms.menu_items_form import MenuItemForm
+from flask_login import login_required, current_user
+from ..models import MenuItem, db
+from ..forms import MenuItemForm
 
-menu_items_routes = Blueprint('menu-items', __name__)
+menu_items_routes = Blueprint("menu-items", __name__)
 
-@menu_items_routes.route('')
+
+@menu_items_routes.route("")
 @login_required
 def show_menu_items():
     menu_items = MenuItem.query.all()
 
-    return {'menu_items': [menu_item.to_dict() for menu_item in menu_items]}
+    return {"menu_items": [menu_item.to_dict() for menu_item in menu_items]}
 
 
-
-@menu_items_routes.route('/<int:id>')
+@menu_items_routes.route("/<int:id>")
 # @login_required
 def item_details(id):
     item_details = MenuItem.query.filter_by(id=id).first()
 
     return {
-            'id': item_details.id,
-            'food_name': item_details.food_name,
-            'description': item_details.description,
-            'price': item_details.price,
-            'image': item_details.img_url
-        }
+        "id": item_details.id,
+        "food_name": item_details.food_name,
+        "description": item_details.description,
+        "price": item_details.price,
+        "image": item_details.img_url,
+    }
+
 
 # moved to restaurant_routes.py
 
@@ -46,7 +47,6 @@ def item_details(id):
 #     return jsonify(message="couldn't addd new item")
 
 
-
 # The bellow routes are defined by Chris Update/Delete
 @menu_items_routes.route("/<int:id>", methods=["PUT"])
 @login_required
@@ -56,6 +56,10 @@ def update_items(id):
 
     if form.validate_on_submit():
         menu_item = MenuItem.query.get(id)
+
+        if int(current_user.get_id()) != menu_item.restaurant.user_id:
+            return {"errors": {"message": "Unauthorized"}}, 401
+
         menu_item.food_name = form.data["food_name"]
         menu_item.decription = form.data["description"]
         menu_item.price = form.data["price"]
@@ -70,6 +74,13 @@ def update_items(id):
 @login_required
 def delete_item(id):
     menu_item = MenuItem.query.get(id)
+
+    if not menu_item:
+        return {"message": "Menu Item couldn't be found"}
+
+    if int(current_user.get_id()) != menu_item.restaurant.user_id:
+        return {"errors": {"message": "Unauthorized"}}, 401
+
     db.session.delete(menu_item)
     db.session.commit()
     return {"message": "Successfully deleted Menu Item"}
